+++ projects/amdsmi/example/amd_smi_drm_example.cc
@@ -1096,6 +1096,11 @@ int main() {
       printf("\tVendor Name: %s\n", asic_info.vendor_name);
       printf("\tSubVendorID: 0x%04x\n", asic_info.subvendor_id);
       printf("\tRevisionID: 0x%02x\n", asic_info.rev_id);
+      if (asic_info.silicon_rev_id != UINT32_MAX) {
+        printf("\tSilicon RevisionID: 0x%02x\n", asic_info.silicon_rev_id);
+      } else {
+        printf("\tSilicon RevisionID: N/A\n");
+      }
       printf("\tSubSystemID: 0x%04x\n", asic_info.subsystem_id);
       printf("\tAsic serial: 0x%s\n", asic_info.asic_serial);
       if (asic_info.oam_id != UINT32_MAX) {