+++ projects/amdsmi/tests/amd_smi_test/test_base.cc
@@ -216,6 +216,12 @@ void TestBase::PrintDeviceHeader(amdsmi_processor_handle dv_ind) {
     }
     std::cout << "\t**Revision ID: 0x" << std::hex << std::setfill('0') << std::setw(2)
               << asic_info.rev_id << std::endl;
+    if (checkIfMaxValue(asic_info.silicon_rev_id)) {
+      std::cout << "\t**Silicon Revision ID: N/A" << std::endl;
+    } else {
+      std::cout << "\t**Silicon Revision ID: 0x" << std::hex << std::setfill('0') << std::setw(2)
+                << asic_info.silicon_rev_id << std::endl;
+    }
     if (checkIfMaxValue(asic_info.subvendor_id)) {
       std::cout << "\t**Subvendor ID: N/A" << std::endl;
     } else {